performance tweaking, splitting the code up
