+++ bench.js
@@ -15,21 +15,34 @@ const martinez = require('./src/index');
  * Asia - JSTS x 6.62 ops/sec ±2.74% (21 runs sampled)
  */
 
+const options = {
+  onStart (event) { console.log(this.name); },
+  onError (event) { console.log(event.target.error); },
+  onCycle (event) { console.log(String(event.target)); },
+  onComplete() {
+    console.log('- Fastest is ' + this.filter('fastest').map('name') + '\n');
+  }
+};
 
+const hole_hole = load.sync('./test/fixtures/hole_hole.geojson')
+new Benchmark.Suite('Hole_Hole', options)
+  .add('Martinez', () => {
+    martinez.union(
+      hole_hole.features[0].geometry.coordinates,
+      hole_hole.features[1].geometry.coordinates);
+  })
+  .add('JSTS', () => {
+    jstsUnion(hole_hole.features[0], hole_hole.features[1]);
+  })
+  .run();
 
-// Define benchmark
-const suite = new Benchmark.Suite('martinez');
-
-var hole_hole = load.sync('./test/fixtures/hole_hole.geojson')
-suite.add('Hole_Hole', () => martinez.union(hole_hole.features[0].geometry.coordinates, hole_hole.features[1].geometry.coordinates));
-suite.add('Hole_Hole - JSTS', () => jstsUnion(hole_hole.features[0], hole_hole.features[1]))
-
-var asia = load.sync('./test/fixtures/asia.geojson')
-var unionPoly = load.sync('./test/fixtures/asia_unionPoly.geojson')
-suite.add('Asia', () => martinez.union(asia.features[0].geometry.coordinates, unionPoly.geometry.coordinates));
-suite.add('Asia - JSTS', () => jstsUnion(asia.features[0], unionPoly))
-
-suite
-  .on('cycle', e => console.log(String(e.target)))
-  .on('complete', () => {})
+const asia = load.sync('./test/fixtures/asia.geojson');
+const unionPoly = load.sync('./test/fixtures/asia_unionPoly.geojson');
+new Benchmark.Suite('Asia union', options)
+  .add('Martinez', () => {
+    martinez.union(
+      asia.features[0].geometry.coordinates,
+      unionPoly.geometry.coordinates);
+  })
+  .add('JSTS', () => jstsUnion(asia.features[0], unionPoly))
   .run();
