+++ demo/index.html
@@ -4,11 +4,15 @@
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="chrome=1">
     <title>Martinez Clipping</title>
-    <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.0.0-rc.2/leaflet.css">
+    <!-- <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.0.0-rc.2/leaflet.css"> -->
+    <link rel="stylesheet" type="text/css" href="../node_modules/leaflet/dist/leaflet.css">
 
-    <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/jsts/1.2.0/jsts.min.js"></script>
-    <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/leaflet/1.0.0-rc.2/leaflet.js"></script>
-    <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/leaflet-editable/0.7.0-beta.1/Leaflet.Editable.min.js"></script>
+    <!-- <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/jsts/1.2.0/jsts.min.js"></script> -->
+    <script type="text/javascript" src="../node_modules/jsts/dist/jsts.min.js"></script>
+    <!-- <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/leaflet/1.0.0-rc.2/leaflet.js"></script> -->
+    <script type="text/javascript" src="../node_modules/leaflet/dist/leaflet.js"></script>
+    <!-- <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/leaflet-editable/0.7.0-beta.1/Leaflet.Editable.min.js"></script> -->
+    <script type="text/javascript" src="../node_modules/leaflet-editable/src/Leaflet.Editable.js"></script>
     <link rel="stylesheet" href="css/styles.css">
     <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
