+++ src/compute_fields.js
@@ -1,0 +1,66 @@
+var edgeType = require('./edge_type');
+
+var INTERSECTION = 0;
+var UNION        = 1;
+var DIFFERENCE   = 2;
+var XOR          = 3;
+
+/**
+ * @param  {SweepEvent} event
+ * @param  {SweepEvent} prev
+ * @param  {Operation} operation
+ */
+module.exports = function computeFields(event, prev, operation) {
+  // compute inOut and otherInOut fields
+  if (prev === null) {
+    event.inOut      = false;
+    event.otherInOut = true;
+
+  // previous line segment in sweepline belongs to the same polygon
+  } else {
+    if (event.isSubject === prev.isSubject) {
+      event.inOut      = !prev.inOut;
+      event.otherInOut = prev.otherInOut;
+
+    // previous line segment in sweepline belongs to the clipping polygon
+    } else {
+      event.inOut      = !prev.otherInOut;
+      event.otherInOut = prev.isVertical() ? !prev.inOut : prev.inOut;
+    }
+
+    // compute prevInResult field
+    if (prev) {
+      event.prevInResult = (!inResult(prev, operation) || prev.isVertical()) ?
+         prev.prevInResult : prev;
+    }
+  }
+
+  // check if the line segment belongs to the Boolean operation
+  event.inResult = inResult(event, operation);
+};
+
+
+function inResult(event, operation) {
+  switch (event.type) {
+  case edgeType.NORMAL:
+    switch (operation) {
+    case INTERSECTION:
+      return !event.otherInOut;
+    case UNION:
+      return event.otherInOut;
+    case DIFFERENCE:
+      return (event.isSubject && event.otherInOut) ||
+              (!event.isSubject && !event.otherInOut);
+    case XOR:
+      return true;
+    }
+    break;
+  case edgeType.SAME_TRANSITION:
+    return operation === INTERSECTION || operation === UNION;
+  case edgeType.DIFFERENT_TRANSITION:
+    return operation === DIFFERENCE;
+  case edgeType.NON_CONTRIBUTING:
+    return false;
+  }
+  return false;
+}
