+++ src/connect_edges.js
@@ -1,0 +1,169 @@
+var equals = require('./equals');
+var compareEvents = require('./compare_events');
+
+
+/**
+ * @param  {Array.<SweepEvent>} sortedEvents
+ * @return {Array.<SweepEvent>}
+ */
+function orderEvents(sortedEvents) {
+  var event, i, len, tmp;
+  var resultEvents = [];
+  for (i = 0, len = sortedEvents.length; i < len; i++) {
+    event = sortedEvents[i];
+    if ((event.left && event.inResult) ||
+      (!event.left && event.otherEvent.inResult)) {
+      resultEvents.push(event);
+    }
+  }
+
+  // Due to overlapping edges the resultEvents array can be not wholly sorted
+  var sorted = false;
+  while (!sorted) {
+    sorted = true;
+    for (i = 0, len = resultEvents.length; i < len; i++) {
+      if ((i + 1) < len &&
+        compareEvents(resultEvents[i], resultEvents[i + 1]) === 1) {
+        tmp = resultEvents[i];
+        resultEvents[i] = resultEvents[i + 1];
+        resultEvents[i + 1] = tmp;
+        sorted = false;
+      }
+    }
+  }
+
+  for (i = 0, len = resultEvents.length; i < len; i++) {
+    resultEvents[i].pos = i;
+  }
+
+  for (i = 0, len = resultEvents.length; i < len; i++) {
+    var event = resultEvents[i];
+    if (!event.left) {
+      tmp = event.pos;
+      event.pos = event.otherEvent.pos;
+      event.otherEvent.pos = tmp;
+    }
+  }
+
+  return resultEvents;
+}
+
+
+/**
+ * @param  {Number} pos
+ * @param  {Array.<SweepEvent>} resultEvents
+ * @param  {Object>}    processed
+ * @return {Number}
+ */
+function nextPos(pos, resultEvents, processed) {
+  var newPos = pos + 1;
+  var length = resultEvents.length;
+  var p  = resultEvents[pos].point;
+  var p1 = resultEvents[newPos].point;
+
+  // if (newPos < length &&
+  //   p1[0] !== p[0] || p1[1] !== p[1] &&
+  //   !processed[newPos]) return newPos;
+  // else return pos - 1;
+
+
+  // while in range and not the current one by value
+  while (newPos < length && p1[0] === p[0] && p1[1] === p[1]) {
+    if (!processed[newPos]) {
+      console.log(pos, newPos, length);
+      return newPos;
+    }
+    else                    newPos++;
+    p1 = resultEvents[newPos].point
+  }
+
+  newPos = pos - 1;
+
+  while (processed[newPos]) newPos--;
+  console.log('other', pos, newPos, length);
+  return newPos;
+}
+
+
+/**
+ * @param  {Array.<SweepEvent>} sortedEvents
+ * @return {Array.<*>} polygons
+ */
+module.exports = function (sortedEvents, operation) {
+  var i, len;
+  var resultEvents = orderEvents(sortedEvents);
+
+  // "false"-filled array
+  var processed = {};
+  var result = [];
+  var event;
+
+  for (i = 0, len = resultEvents.length; i < len; i++) {
+    if (processed[i]) continue;
+    var contour = [[]];
+
+    if (!resultEvents[i].isExteriorRing) {
+      if (result.length === 0) {
+        result.push([[contour]]);
+      } else {
+        if (operation === 1) {
+          result.push(contour)
+        } else {
+          result[result.length - 1].push(contour);
+        }
+      }
+    } else {
+      result.push(contour);
+    }
+
+    var ringId = result.length - 1;
+    var pos = i;
+
+    var initial = resultEvents[i].point;
+    // initial.push(resultEvents[i].isExteriorRing);
+    contour[0].push(initial);
+
+    while (pos >= i) {
+      var event = resultEvents[pos];
+      processed[pos] = true;
+
+      if (event.left) {
+        event.resultInOut = false;
+        event.contourId   = ringId;
+      } else {
+        event.otherEvent.resultInOut = true;
+        event.otherEvent.contourId   = ringId;
+      }
+
+      pos = event.pos;
+      processed[pos] = true;
+      // resultEvents[pos].point.push(resultEvents[pos].isExteriorRing);
+
+      contour[0].push(resultEvents[pos].point);
+      pos = nextPos(pos, resultEvents, processed);
+    }
+
+    pos = pos === -1 ? i : pos;
+
+    event = resultEvents[pos];
+    processed[pos] = processed[event.pos] = true;
+    event.otherEvent.resultInOut = true;
+    event.otherEvent.contourId   = ringId;
+  }
+
+  for (i = 0, len = result.length; i < len; i++) {
+    var polygon = result[i];
+    for (var j = 0, jj = polygon.length; j < jj; j++) {
+      var polygonContour = polygon[j];
+      for (var k = 0, kk = polygonContour.length; k < kk; k++) {
+        var coords = polygonContour[k];
+        if (typeof coords[0] !== 'number') {
+          polygon.push(coords[0]);
+          polygon.splice(j, 1);
+        }
+      }
+    }
+  }
+
+  return result;
+};
