+++ src/fill_queue.js
@@ -1,0 +1,70 @@
+var Queue           = require('tinyqueue');
+var SweepEvent      = require('./sweep_event');
+var compareEvents   = require('./compare_events');
+
+var max = Math.max;
+var min = Math.min;
+
+var contourId = 0;
+
+
+function processPolygon(contourOrHole, isSubject, depth, Q, bbox, isExteriorRing) {
+  var i, len, s1, s2, e1, e2;
+  var d = depth + 1;
+  for (i = 0, len = contourOrHole.length - 1; i < len; i++) {
+    s1 = contourOrHole[i],
+    s2 = contourOrHole[i + 1];
+    //processSegment(contourOrHole[i], contourOrHole[i + 1], isSubject, depth + 1, Q, bbox, isExteriorRing);
+    e1 = new SweepEvent(s1, false, undefined, isSubject);
+    e2 = new SweepEvent(s2, false, e1,        isSubject);
+    e1.otherEvent = e2;
+
+    e1.contourId = e2.contourId = depth;
+    if (!isExteriorRing) {
+      e1.isExteriorRing = false;
+      e2.isExteriorRing = false;
+    }
+    if (compareEvents(e1, e2) > 0) {
+      e2.left = true;
+    } else {
+      e1.left = true;
+    }
+
+    var x = s1[0], y = s1[1];
+    bbox[0] = min(bbox[0], x);
+    bbox[1] = min(bbox[1], y);
+    bbox[2] = max(bbox[2], x);
+    bbox[3] = max(bbox[3], y);
+
+    // Pushing it so the queue is sorted from left to right,
+    // with object on the left having the highest priority.
+    Q.push(e1);
+    Q.push(e2);
+  }
+}
+
+
+module.exports = function fillQueue(subject, clipping, sbbox, cbbox) {
+  var eventQueue = new Queue(null, compareEvents);
+  var polygonSet, isExteriorRing, i, ii, j, jj, k, kk;
+
+  for (i = 0, ii = subject.length; i < ii; i++) {
+    polygonSet = subject[i];
+    for (j = 0, jj = polygonSet.length; j < jj; j++) {
+      isExteriorRing = j === 0;
+      if (isExteriorRing) contourId++;
+      processPolygon(polygonSet[j], true, contourId, eventQueue, sbbox, isExteriorRing);
+    }
+  }
+
+  for (i = 0, ii = clipping.length; i < ii; i++) {
+    polygonSet = clipping[i];
+    for (j = 0, jj = polygonSet.length; j < jj; j++) {
+      isExteriorRing = j === 0;
+      if (isExteriorRing) contourId++;
+      processPolygon(polygonSet[j], false, contourId, eventQueue, cbbox, isExteriorRing);
+    }
+  }
+
+  return eventQueue;
+}
