+++ src/operation.js
@@ -1,0 +1,6 @@
+module.exports = {
+  INTERSECTION: 0,
+  UNION:        1,
+  DIFFERENCE:   2,
+  XOR:          3
+};
