+++ src/possible_intersection.js
@@ -1,0 +1,109 @@
+var divideSegment = require('./divide_segment');
+var intersection  = require('./segment_intersection');
+var equals        = require('./equals');
+
+/**
+ * @param  {SweepEvent} se1
+ * @param  {SweepEvent} se2
+ * @param  {Queue}      queue
+ * @return {Number}
+ */
+module.exports = function possibleIntersection(se1, se2, queue) {
+  // that disallows self-intersecting polygons,
+  // did cost us half a day, so I'll leave it
+  // out of respect
+  // if (se1.isSubject === se2.isSubject) return;
+  var inter = intersection(
+    se1.point, se1.otherEvent.point,
+    se2.point, se2.otherEvent.point
+  );
+
+  var nintersections = inter ? inter.length : 0;
+  if (nintersections === 0) return 0; // no intersection
+
+  // the line segments intersect at an endpoint of both line segments
+  if ((nintersections === 1) &&
+      (equals(se1.point, se2.point) ||
+       equals(se1.otherEvent.point, se2.otherEvent.point))) {
+    return 0;
+  }
+
+  if (nintersections === 2 && se1.isSubject === se2.isSubject) {
+    // if(se1.contourId === se2.contourId){
+    // console.warn('Edges of the same polygon overlap',
+    //   se1.point, se1.otherEvent.point, se2.point, se2.otherEvent.point);
+    // }
+    //throw new Error('Edges of the same polygon overlap');
+    return 0;
+  }
+
+  // The line segments associated to se1 and se2 intersect
+  if (nintersections === 1) {
+
+    // if the intersection point is not an endpoint of se1
+    if (!equals(se1.point, inter[0]) && !equals(se1.otherEvent.point, inter[0])) {
+      divideSegment(se1, inter[0], queue);
+    }
+
+    // if the intersection point is not an endpoint of se2
+    if (!equals(se2.point, inter[0]) && !equals(se2.otherEvent.point, inter[0])) {
+      divideSegment(se2, inter[0], queue);
+    }
+    return 1;
+  }
+
+  // The line segments associated to se1 and se2 overlap
+  var events        = [];
+  var leftCoincide  = false;
+  var rightCoincide = false;
+
+  if (equals(se1.point, se2.point)) {
+    leftCoincide = true; // linked
+  } else if (compareEvents(se1, se2) === 1) {
+    events.push(se2, se1);
+  } else {
+    events.push(se1, se2);
+  }
+
+  if (equals(se1.otherEvent.point, se2.otherEvent.point)) {
+    rightCoincide = true;
+  } else if (compareEvents(se1.otherEvent, se2.otherEvent) === 1) {
+    events.push(se2.otherEvent, se1.otherEvent);
+  } else {
+    events.push(se1.otherEvent, se2.otherEvent);
+  }
+
+  if ((leftCoincide && rightCoincide) || leftCoincide) {
+    // both line segments are equal or share the left endpoint
+    se1.type = edgeType.NON_CONTRIBUTING;
+    se2.type = (se1.inOut === se2.inOut) ?
+      edgeType.SAME_TRANSITION :
+      edgeType.DIFFERENT_TRANSITION;
+
+    if (leftCoincide && !rightCoincide) {
+      // honestly no idea, but changing events selection from [2, 1]
+      // to [0, 1] fixes the overlapping self-intersecting polygons issue
+      divideSegment(events[1].otherEvent, events[0].point, queue);
+    }
+    return 2;
+  }
+
+  // the line segments share the right endpoint
+  if (rightCoincide) {
+    divideSegment(events[0], events[1].point, queue);
+    return 3;
+  }
+
+  // no line segment includes totally the other one
+  if (events[0] !== events[3].otherEvent) {
+    divideSegment(events[0], events[1].point, queue);
+    divideSegment(events[1], events[2].point, queue);
+    return 3;
+  }
+
+  // one line segment includes the other one
+  divideSegment(events[0], events[1].point, queue);
+  divideSegment(events[3].otherEvent, events[2].point, queue);
+
+  return 3;
+};
