+++ src/subdivide_segments.js
@@ -1,0 +1,118 @@
+var Tree                 = require('avl');
+var computeFields        = require('./compute_fields');
+var possibleIntersection = require('./possible_intersection');
+var compareSegments      = require('./compare_segments');
+var operations           = require('./operation');
+
+
+module.exports = function subdivide (eventQueue, subject, clipping, sbbox, cbbox, operation) {
+  var sweepLine = new Tree(compareSegments);
+  var sortedEvents = [];
+
+  var rightbound = Math.min(sbbox[2], cbbox[2]);
+
+  var prev, next, begin, end;
+
+  var INTERSECTION = operations.INTERSECTION;
+  var DIFFERENCE   = operations.DIFFERENCE;
+
+  while (eventQueue.length) {
+    var event = eventQueue.pop();
+    sortedEvents.push(event);
+
+    // optimization by bboxes for intersection and difference goes here
+    if ((operation === INTERSECTION && event.point[0] > rightbound) ||
+        (operation === DIFFERENCE   && event.point[0] > sbbox[2])) {
+      break;
+    }
+
+    if (event.left) {
+      next  = prev = sweepLine.insert(event);
+      //_renderSweepLine(sweepLine, event.point, event);
+      begin = sweepLine.minNode();
+
+      if (prev !== begin) prev = sweepLine.prev(prev);
+      else                prev = null;
+
+      next = sweepLine.next(next);
+
+      var prevEvent = prev ? prev.key : null;
+      var prevprevEvent;
+      computeFields(event, prevEvent, operation);
+      if (next) {
+        if (possibleIntersection(event, next.key, eventQueue) === 2) {
+          computeFields(event, prevEvent, operation);
+          computeFields(event, next.key, operation);
+        }
+      }
+      if (prev) {
+        if (possibleIntersection(prev.key, event, eventQueue) === 2) {
+          var prevprev = prev;
+          if (prevprev !== begin) prevprev = sweepLine.prev(prevprev);
+          else                    prevprev = null;
+
+          prevprevEvent = prevprev ? prevprev.key : null;
+          computeFields(prevEvent, prevprevEvent, operation);
+          computeFields(event,     prevEvent,     operation);
+        }
+      }
+    } else {
+      event = event.otherEvent;
+      next = prev = sweepLine.find(event);
+
+      // _renderSweepLine(sweepLine, event.otherEvent.point, event);
+
+      if (prev && next) {
+
+        if (prev !== begin) prev = sweepLine.prev(prev);
+        else                prev = null;
+
+        next = sweepLine.next(next);
+        sweepLine.remove(event);
+
+        // _renderSweepLine(sweepLine, event.otherEvent.point, event);
+
+        if (next && prev) {
+          //if (typeof prev !== 'undefined' && typeof next !== 'undefined') {
+          possibleIntersection(prev.key, next.key, eventQueue);
+          //}
+        }
+      }
+    }
+  }
+  return sortedEvents;
+};
+
+
+/* eslint-disable no-unused-vars, no-debugger, no-undef */
+function _renderSweepLine(sweepLine, pos, event) {
+  var map = window.map;
+  if (!map) return;
+  if (window.sws) window.sws.forEach(function (p) {
+    map.removeLayer(p);
+  });
+  window.sws = [];
+  sweepLine.forEach(function (e) {
+    var poly = L.polyline([
+      e.point.slice().reverse(),
+      e.otherEvent.point.slice().reverse()
+    ], {color: 'green'}).addTo(map);
+    window.sws.push(poly);
+  });
+
+  if (window.vt) map.removeLayer(window.vt);
+  var v = pos.slice();
+  var b = map.getBounds();
+  window.vt = L.polyline([
+    [b.getNorth(), v[0]],
+    [b.getSouth(), v[0]]
+  ], {color: 'green', weight: 1}).addTo(map);
+
+  if (window.ps) map.removeLayer(window.ps);
+  window.ps = L.polyline([
+    event.point.slice().reverse(),
+    event.otherEvent.point.slice().reverse()
+  ], {color: 'black', weight: 9, opacity: 0.4}).addTo(map);
+  debugger;
+}
+/* eslint-enable no-unused-vars, no-debugger, no-undef */
